Feature ETP-4329: Rename InvoiceCobroModal to InvoicePaymentHistoryModal
