+++ artifacts/sales-invoice/custom/InvoiceHeaderTable.jsx
@@ -13,7 +13,7 @@ import {
 import { useFiscalConfig } from '@/windows/custom/fiscal-config/useFiscalConfig.js';
 import { getInvoiceFiscalTargets } from '@/windows/custom/shared/fiscalTargets.js';
 import { FiscalStatusBadge, normalizeVerifactuStatus } from '@/windows/custom/shared/FiscalStatusBadge.jsx';
-import InvoiceCobroModal from '@/windows/custom/shared/InvoiceCobroModal.jsx';
+import InvoicePaymentHistoryModal from '@/windows/custom/shared/InvoicePaymentHistoryModal.jsx';
 import { getArSubtype } from './invoiceSubtype';
 
 // ─── Invoice-specific status logic ───────────────────────────────
@@ -181,7 +181,7 @@ export default function InvoiceHeaderTable(props) {
     <>
       <DataTable columns={columns} filters={FILTERS} {...props} />
       {paymentRow && (
-        <InvoiceCobroModal
+        <InvoicePaymentHistoryModal
           invoiceId={paymentRow.id}
           invoiceData={paymentRow}
           specName="sales-invoice"
